#2133 reorganized pages
